# api/openapi.yaml
openapi: 3.0.3
info:
  title: Changeloom API
  version: 1.0.0
  description: >
    Personal tech-news timeline. All /v1 endpoints require a Firebase ID token
    in `Authorization: Bearer <token>`. Requests are rate limited per client IP
    and per user; over the limit the api answers 429 with a Retry-After header.
    Responses under /v1 are `Cache-Control: private, no-store` unless an operation
    says otherwise.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /v1/topic-requests:
    get:
      operationId: listMyTopicRequests
      summary: The signed-in user's topic requests, newest first.
      responses:
        "200":
          description: The user's topic requests.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/TopicRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
    post:
      operationId: createTopicRequest
      summary: Ask for a topic to be added to the catalog.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: {type: string, description: "What to follow, 2-100 characters after trimming."}
      responses:
        "201":
          description: The created request.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TopicRequest"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "409":
          description: The same request is already pending.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
        "429":
          description: Too many pending requests.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Error"}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
...
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
    Unauthorized:
      description: Missing or invalid token.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: {type: string, example: not_found}
        message: {type: string}
...
    TopicRequest:
      type: object
      required: [id, text, status, created_at]
      properties:
        id: {type: integer, format: int64}
        text: {type: string}
        status:
          type: string
          enum: [pending, accepted, merged, rejected]
          description: accepted means a new topic was created; merged means it maps to an existing one.
        topic: {type: string, description: Slug of the topic the request was resolved to.}
        note: {type: string, description: Reviewer note.}
        created_at: {type: string, format: date-time}
        resolved_at: {type: string, format: date-time}
...